Add description to categories:

+++ db/migrate/20221227235655_create_user_bills.rb
@@ -3,7 +3,7 @@ class CreateUserBills < ActiveRecord::Migration[6.1]
     create_table :user_bills do |t|
       t.string :name
       t.string :description
-      t.integer :amount
+      t.decimal :amount
       t.references :user, foreign_key: true, null: false
       t.references :expense_category, foreign_key: true, null: false
